requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
websocket-client==0.37.0
